+++ forelesninger/F14/src/no/hvl/dat100/unntak/eksempler/HastighetException.java
@@ -13,7 +13,7 @@ public class HastighetException {
 			String tmstr = JOptionPane.showInputDialog("Tid i timer");
 			int tm = Integer.parseInt(tmstr); 
 
-			double hastighet = (double)km/tm;
+			int hastighet = km/tm;
 			System.out.println("Hastighet: " + hastighet);
 			
 		} catch (NumberFormatException e) {
